+++ packages/example/webpack/webpack.prod.js
@@ -13,7 +13,6 @@ module.exports = merge(common, {
     minimize: true,
     minimizer: [new TerserPlugin({})],
   },
-
   plugins: [
     new HtmlWebpackPlugin({
       inject: true,